requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
alembic
databases
pyjwt
passlib[bcrypt]